cleanup: Get rid of binary deps and build artifacts
Build artifacts belong in the binary distribution, not the source distribution. They will inevitably go out of sync with the code, causing more trouble than they are worth. That's before considering that there may be as many artifacts as users.

Binary dependencies belong in the binary distribution, or maybe even that's the distro's problem. They will go stale the moment the ABI changes, and they will make it harder to justify expanding to new architectures.

Instead there should be a good documentation and automated tests that can retrieve those binaries whenever needed.