build: Embed git revision in version in a sane(r) way
The previous way of embedding git revision was to count the number of commits leading to the current one. This is bad because with branching, there can exist two sibling commits with the same number of ancestors and therefore same version. Confusion would arise.

After the change, the revision is included directly in the version string.

That is not enough to ensure two binaries with the same version string are built from the same source, because git directories are often work in progress. To make possible to see if the files haven't been altered, the "-dirty" flag is added if they were.

That's still short of reproducible builds, but should get us a long way.